trainer/requirements.txt
datasets==3.6.0
dp_transformers==1.0.0
flash_attn==2.8.3
numpy==2.4.1
opacus==1.5.4
peft==0.17.1
prv_accountant==0.2.0
PyYAML==6.0.2
torch==2.8.0
transformers==4.56.1
xformers==0.0.34
